+++ src/index.js
@@ -1,50 +1,82 @@
 (function(){
-  var chise = window.chise = function(_options, _libs) {
+  var chise = function(_options) {
+
+    var param = {};
+
+    // Access the libs
+    var libs = require('./utilities/lib-utilities').getLibs();
+
+    var optionUtilities = require('./utilities/option-utilities-factory')();
+    var options = optionUtilities.extendOptions(_options); // Extends the default options with the given options
+
+    // Create an sbgnviz instance
+    var sbgnvizInstanceUtilities = require('./utilities/sbgnviz-instance-utilities-factory')();
+    var sbgnvizInstance = sbgnvizInstanceUtilities(options);
+
+    // Update style and bind events
+    var cyStyleAndEvents = require('./utilities/cy-style-and-events-factory')();
+
+    // Register undo/redo actions
+    var registerUndoRedoActions = require('./utilities/register-undo-redo-actions-factory')();
+
+    var mainUtilities = require('./utilities/main-utilities-factory')();
+    var elementUtilitiesExtender = require('./utilities/element-utilities-extender-factory')();
+    var undoRedoActionFunctionsExtender = require('./utilities/ur-action-functions-extender-factory')();
+
+    var elementUtilities =  sbgnvizInstance.elementUtilities;
+    var undoRedoActionFunctions = sbgnvizInstance.undoRedoActionFunctions;
+
+    param.sbgnvizInstanceUtilities = sbgnvizInstanceUtilities;
+    param.optionUtilities = optionUtilities;
+    param.elementUtilities = elementUtilities;
+    param.undoRedoActionFunctions = undoRedoActionFunctions;
+
+    undoRedoActionFunctionsExtender(param);
+    elementUtilitiesExtender(param);
+    cyStyleAndEvents(param);
+    registerUndoRedoActions(param);
+    mainUtilities(param);
+
+    // Expose the api
+    var api = {};
+
+    // Expose the properties inherited from sbgnviz
+    // then override some of these properties and expose some new properties
+    for (var prop in sbgnvizInstance) {
+      api[prop] = sbgnvizInstance[prop];
+    }
+
+    // Expose each main utility seperately
+    for (var prop in mainUtilities) {
+      api[prop] = mainUtilities[prop];
+    }
+
+    // Expose getSbgnvizInstance()
+    api.getSbgnvizInstance = sbgnvizInstanceUtilities.getSbgnvizInstance;
+
+    // Expose elementUtilities and undoRedoActionFunctions as is
+    api.elementUtilities = elementUtilities;
+    api.undoRedoActionFunctions = undoRedoActionFunctions;
+
+    return api;
+  };
+
+  // Register chise with given libraries
+  chise.register = function (_libs) {
+
     var libs = {};
     libs.jQuery = _libs.jQuery || jQuery;
     libs.cytoscape = _libs.cytoscape || cytoscape;
     libs.sbgnviz = _libs.sbgnviz || sbgnviz;
     libs.saveAs = _libs.filesaverjs ? _libs.filesaverjs.saveAs : saveAs;
-    
-    libs.sbgnviz(_options, _libs); // Initilize sbgnviz
-    
+
+    libs.sbgnviz.register(_libs); // Register sbgnviz with the given libs
+
     // Set the libraries to access them from any file
     var libUtilities = require('./utilities/lib-utilities');
     libUtilities.setLibs(libs);
-    
-    var optionUtilities = require('./utilities/option-utilities');
-    var options = optionUtilities.extendOptions(_options); // Extends the default options with the given options
-    
-    // Update style and bind events
-    var cyStyleAndEvents = require('./utilities/cy-style-and-events');
-    cyStyleAndEvents(libs.sbgnviz);
-    
-    // Register undo/redo actions
-    var registerUndoRedoActions = require('./utilities/register-undo-redo-actions');
-    registerUndoRedoActions(options.undoableDrag);
-    
-    var mainUtilities = require('./utilities/main-utilities');
-    var elementUtilities = require('./utilities/element-utilities');
-    var undoRedoActionFunctions = require('./utilities/undo-redo-action-functions');
-    
-    // Expose the api
-    
-    // Expose the properties inherited from sbgnviz
-    // then override some of these properties and expose some new properties
-    for (var prop in libs.sbgnviz) {
-      chise[prop] = libs.sbgnviz[prop];
-    }
-    
-    // Expose each main utility seperately
-    for (var prop in mainUtilities) {
-      chise[prop] = mainUtilities[prop];
-    }
-    
-    // Expose elementUtilities and undoRedoActionFunctions as is
-    chise.elementUtilities = elementUtilities;
-    chise.undoRedoActionFunctions = undoRedoActionFunctions;
   };
-  
+
   if ( typeof module !== 'undefined' && module.exports ) {
     module.exports = chise;
   }
