+++ src/utilities/main-utilities-factory.js
@@ -1,0 +1,937 @@
+var libs = require('./lib-utilities').getLibs();
+
+/*
+ * The main utilities to be exposed directly.
+ */
+module.exports = function () {
+
+  var elementUtilities, options, cy;
+
+  function mainUtilities (param) {
+    elementUtilities = param.elementUtilities;
+    options = param.optionUtilities.getOptions();
+    cy = param.sbgnvizInstanceUtilities.getCy();
+  };
+
+  /*
+   * Adds a new node with the given class and at the given coordinates. Considers undoable option.
+   */
+  mainUtilities.addNode = function(x, y , nodeParams, id, parent, visibility) {
+    // update map type
+    if (typeof nodeParams == 'object'){
+
+      if (!elementUtilities.getMapType())
+        elementUtilities.setMapType(nodeParams.language);
+      else if (elementUtilities.getMapType() != nodeParams.language)
+        elementUtilities.setMapType("Unknown");
+    }
+
+    if (!options.undoable) {
+      return elementUtilities.addNode(x, y, nodeParams, id, parent, visibility);
+    }
+    else {
+      var param = {
+        newNode : {
+          x: x,
+          y: y,
+          class: nodeParams,
+          id: id,
+          parent: parent,
+          visibility: visibility
+        }
+      };
+
+      cy.undoRedo().do("addNode", param);
+    }
+  };
+
+  /*
+   * Adds a new edge with the given class and having the given source and target ids. Considers undoable option.
+   */
+  mainUtilities.addEdge = function(source, target, edgeParams, id, visibility) {
+    // update map type
+    if (typeof edgeParams == 'object'){
+
+      if (!elementUtilities.getMapType())
+        elementUtilities.setMapType(edgeParams.language);
+      else if (elementUtilities.getMapType() != edgeParams.language)
+        elementUtilities.setMapType("Unknown");
+    }
+    // Get the validation result
+    var edgeclass = edgeParams.class ? edgeParams.class : edgeParams;
+    var validation = elementUtilities.validateArrowEnds(edgeclass, cy.getElementById(source), cy.getElementById(target));
+
+    // If validation result is 'invalid' cancel the operation
+    if (validation === 'invalid') {
+      return;
+    }
+
+    // If validation result is 'reverse' reverse the source-target pair before creating the edge
+    if (validation === 'reverse') {
+      var temp = source;
+      source = target;
+      target = temp;
+    }
+
+    if (!options.undoable) {
+      return elementUtilities.addEdge(source, target, edgeParams, id, visibility);
+    }
+    else {
+      var param = {
+        newEdge : {
+          source: source,
+          target: target,
+          class: edgeParams,
+          id: id,
+          visibility: visibility
+        }
+      };
+
+      cy.undoRedo().do("addEdge", param);
+    }
+  };
+
+  /*
+   * Adds a process with convenient edges. For more information please see 'https://github.com/iVis-at-Bilkent/newt/issues/9'.
+   * Considers undoable option.
+   */
+  mainUtilities.addProcessWithConvenientEdges = function(_source, _target, processType) {
+    // If source and target IDs are given get the elements by IDs
+    var source = typeof _source === 'string' ? cy.getElementById(_source) : _source;
+    var target = typeof _target === 'string' ? cy.getElementById(_target) : _target;
+
+    // If source or target does not have an EPN class the operation is not valid
+    if (!elementUtilities.isEPNClass(source) || !elementUtilities.isEPNClass(target)) {
+      return;
+    }
+
+    if (!options.undoable) {
+      return elementUtilities.addProcessWithConvenientEdges(_source, _target, processType);
+    }
+    else {
+      var param = {
+        source: _source,
+        target: _target,
+        processType: processType
+      };
+
+      cy.undoRedo().do("addProcessWithConvenientEdges", param);
+    }
+  };
+
+  // convert collapsed compound nodes to simple nodes
+  // and update port values of pasted nodes and edges
+  var cloneCollapsedNodesAndPorts = function (elesBefore){
+    cy.elements().unselect();
+    var elesAfter = cy.elements();
+    var elesDiff = elesAfter.diff(elesBefore).left;
+
+    // shallow copy collapsed nodes - collapsed compounds become simple nodes
+    // data related to collapsed nodes are removed from generated clones
+    // related issue: https://github.com/iVis-at-Bilkent/newt/issues/145
+    var collapsedNodes = elesDiff.filter('node.cy-expand-collapse-collapsed-node');
+
+    collapsedNodes.connectedEdges().remove();
+    collapsedNodes.removeClass('cy-expand-collapse-collapsed-node');
+    collapsedNodes.removeData('collapsedChildren');
+    collapsedNodes.removeData('position-before-collapse size-before-collapse');
+    collapsedNodes.removeData('expandcollapseRenderedCueSize expandcollapseRenderedStartX expandcollapseRenderedStartY');
+
+    // cloning ports
+    elesDiff.nodes().forEach(function(_node){
+      if(_node.data("ports").length == 2){
+          var oldPortName0 = _node.data("ports")[0].id;
+          var oldPortName1 = _node.data("ports")[1].id;
+          _node.data("ports")[0].id = _node.id() + ".1";
+          _node.data("ports")[1].id = _node.id() + ".2";
+
+          _node.outgoers().edges().forEach(function(_edge){
+            if(_edge.data("portsource") == oldPortName0){
+              _edge.data("portsource", _node.data("ports")[0].id);
+            }
+            else if(_edge.data("portsource") == oldPortName1){
+              _edge.data("portsource", _node.data("ports")[1].id);
+            }
+            else{
+              _edge.data("portsource", _node.id());
+            }
+          });
+          _node.incomers().edges().forEach(function(_edge){
+            if(_edge.data("porttarget") == oldPortName0){
+              _edge.data("porttarget", _node.data("ports")[0].id);
+            }
+            else if(_edge.data("porttarget") == oldPortName1){
+              _edge.data("porttarget", _node.data("ports")[1].id);
+            }
+            else{
+              _edge.data("porttarget", _node.id());
+            }
+          });
+      }
+      else{
+        _node.outgoers().edges().forEach(function(_edge){
+          _edge.data("portsource", _node.id());
+        });
+        _node.incomers().edges().forEach(function(_edge){
+          _edge.data("porttarget", _node.id());
+        });
+      }
+    });
+    elesDiff.select();
+  }
+
+  /*
+   * Clone given elements. Considers undoable option. Requires cytoscape-clipboard extension.
+   */
+  mainUtilities.cloneElements = function (eles) {
+    if (eles.length === 0) {
+      return;
+    }
+    var elesBefore = cy.elements();
+
+    var cb = cy.clipboard();
+    var _id = cb.copy(eles, "cloneOperation");
+
+    if (options.undoable) {
+      cy.undoRedo().do("paste", {id: _id});
+    }
+    else {
+      cb.paste(_id);
+    }
+    cloneCollapsedNodesAndPorts(elesBefore);
+  };
+
+  /*
+   * Copy given elements to clipboard. Requires cytoscape-clipboard extension.
+   */
+  mainUtilities.copyElements = function (eles) {
+    cy.clipboard().copy(eles);
+  };
+
+  /*
+   * Past the elements copied to clipboard. Considers undoable option. Requires cytoscape-clipboard extension.
+   */
+  mainUtilities.pasteElements = function() {
+    var elesBefore = cy.elements();
+
+    if (options.undoable) {
+      cy.undoRedo().do("paste");
+    }
+    else {
+      cy.clipboard().paste();
+    }
+    cloneCollapsedNodesAndPorts(elesBefore);
+  };
+
+  /*
+   * Aligns given nodes in given horizontal and vertical order.
+   * Horizontal and vertical parameters may be 'none' or undefined.
+   * alignTo parameter indicates the leading node.
+   * Requrires cytoscape-grid-guide extension and considers undoable option.
+   */
+  mainUtilities.align = function (nodes, horizontal, vertical, alignTo) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (options.undoable) {
+      cy.undoRedo().do("align", {
+        nodes: nodes,
+        horizontal: horizontal,
+        vertical: vertical,
+        alignTo: alignTo
+      });
+    } else {
+      nodes.align(horizontal, vertical, alignTo);
+    }
+  };
+
+  /*
+   * Create compound for given nodes. compoundType may be 'complex' or 'compartment'.
+   * This method considers undoable option.
+   */
+  mainUtilities.createCompoundForGivenNodes = function (_nodes, compoundType) {
+    var nodes = _nodes;
+    /*
+     * Eleminate the nodes which cannot have a parent with given compound type
+     */
+    nodes = _nodes.filter(function (element, i) {
+      if(typeof element === "number") {
+        element = i;
+      }
+
+      var sbgnclass = element.data("class");
+      return elementUtilities.isValidParent(sbgnclass, compoundType, element);
+    });
+
+    nodes = elementUtilities.getTopMostNodes(nodes);
+
+    // All elements should have the same parent and the common parent should not be a 'complex'
+    // if compoundType is 'compartent'
+    // because the old common parent will be the parent of the new compartment after this operation and
+    // 'complexes' cannot include 'compartments'
+    if (nodes.length == 0 || !elementUtilities.allHaveTheSameParent(nodes)
+            || ( (compoundType === 'compartment' || compoundType == 'submap') && nodes.parent().data('class')
+            && nodes.parent().data('class').startsWith('complex') )) {
+      return;
+    }
+
+    if (cy.undoRedo()) {
+      var param = {
+        compoundType: compoundType,
+        nodesToMakeCompound: nodes
+      };
+
+      cy.undoRedo().do("createCompoundForGivenNodes", param);
+    }
+    else {
+      elementUtilities.createCompoundForGivenNodes(nodes, compoundType);
+    }
+  };
+
+  /*
+   * Move the nodes to a new parent and change their position if possDiff params are set.
+   * Considers undoable option and checks if the operation is valid.
+   */
+  mainUtilities.changeParent = function(nodes, _newParent, posDiffX, posDiffY) {
+    var newParent = typeof _newParent === 'string' ? cy.getElementById(_newParent) : _newParent;
+    // New parent is supposed to be one of the root, a complex or a compartment
+    if (newParent && !newParent.data("class").startsWith("complex") && newParent.data("class") != "compartment"
+            && newParent.data("class") != "submap") {
+      return;
+    }
+    /*
+     * Eleminate the nodes which cannot have the newParent as their parent
+     */
+    nodes = nodes.filter(function (element, i) {
+      if(typeof element === "number") {
+        element = i;
+      }
+
+      var sbgnclass = element.data("class");
+      return elementUtilities.isValidParent(sbgnclass, newParent, element);
+    });
+
+    // Discard the nodes whose parent is already newParent.
+    // Discard the newParent itself if it is among the nodes
+    nodes = nodes.filter(function (ele, i) {
+      if(typeof ele === "number") {
+        ele = i;
+      }
+
+      // Discard the newParent if it is among the nodes
+      if (newParent && ele.id() === newParent.id()) {
+        return false;
+      }
+      // Discard the nodes whose parent is already newParent
+      if (!newParent) {
+        return ele.data('parent') != null;
+      }
+      return ele.data('parent') !== newParent.id();
+    });
+
+    // If some nodes are ancestor of new parent eleminate them
+    if (newParent) {
+      nodes = nodes.difference(newParent.ancestors());
+    }
+
+    // If all nodes are eleminated return directly
+    if (nodes.length === 0) {
+      return;
+    }
+
+    // Just move the top most nodes
+    nodes = elementUtilities.getTopMostNodes(nodes);
+
+    var parentId = newParent ? newParent.id() : null;
+
+    if (options.undoable) {
+      var param = {
+        firstTime: true,
+        parentData: parentId, // It keeps the newParentId (Just an id for each nodes for the first time)
+        nodes: nodes,
+        posDiffX: posDiffX,
+        posDiffY: posDiffY,
+        // This is needed because the changeParent function called is not from elementUtilities
+        // but from the undoRedo extension directly, so maintaining pointer is not automatically done.
+        callback: elementUtilities.maintainPointer
+      };
+
+      cy.undoRedo().do("changeParent", param); // This action is registered by undoRedo extension
+    }
+    else {
+      elementUtilities.changeParent(nodes, parentId, posDiffX, posDiffY);
+    }
+  };
+
+  /*
+   * Creates a template reaction with given parameters. Requires cose-bilkent layout to tile the free macromolecules included
+   * in the complex. Considers undoable option. For more information see the same function in elementUtilities
+   */
+  mainUtilities.createTemplateReaction = function (templateType, macromoleculeList, complexName, processPosition, tilingPaddingVertical, tilingPaddingHorizontal, edgeLength) {
+    if (!options.undoable) {
+      elementUtilities.createTemplateReaction(templateType, macromoleculeList, complexName, processPosition, tilingPaddingVertical, tilingPaddingHorizontal, edgeLength);
+    }
+    else {
+      var param = {
+        templateType: templateType,
+        macromoleculeList: macromoleculeList,
+        complexName: complexName,
+        processPosition: processPosition,
+        tilingPaddingVertical: tilingPaddingVertical,
+        tilingPaddingHorizontal: tilingPaddingHorizontal,
+        edgeLength: edgeLength
+      };
+
+      cy.undoRedo().do("createTemplateReaction", param);
+    }
+  };
+
+  /*
+   * Resize given nodes if useAspectRatio is truthy one of width or height should not be set.
+   * Considers undoable option.
+   */
+  mainUtilities.resizeNodes = function(nodes, width, height, useAspectRatio) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (options.undoable) {
+      var param = {
+        nodes: nodes,
+        width: width,
+        height: height,
+        useAspectRatio: useAspectRatio,
+        performOperation: true
+      };
+
+      cy.undoRedo().do("resizeNodes", param);
+    }
+    else {
+      elementUtilities.resizeNodes(nodes, width, height, useAspectRatio);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Changes the label of the given nodes to the given label. Considers undoable option.
+   */
+  mainUtilities.changeNodeLabel = function(nodes, label) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (!options.undoable) {
+      nodes.data('label', label);
+    }
+    else {
+      var param = {
+        nodes: nodes,
+        label: label,
+        firstTime: true
+      };
+
+      cy.undoRedo().do("changeNodeLabel", param);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Change font properties for given nodes use the given font data.
+   * Considers undoable option.
+   */
+  mainUtilities.changeFontProperties = function(eles, data) {
+    if (eles.length === 0) {
+      return;
+    }
+
+    if (options.undoable) {
+      var param = {
+        eles: eles,
+        data: data,
+        firstTime: true
+      };
+
+      cy.undoRedo().do("changeFontProperties", param);
+    }
+    else {
+      elementUtilities.changeFontProperties(eles, data);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Change state value or unit of information box of given nodes with given index.
+   * Considers undoable option.
+   * For more information about the parameters see elementUtilities.changeStateOrInfoBox
+   */
+  mainUtilities.changeStateOrInfoBox = function(nodes, index, value, type) {
+    if (nodes.length === 0) {
+      return;
+    }
+    if (options.undoable) {
+      var param = {
+        index: index,
+        value: value,
+        type: type,
+        nodes: nodes
+      };
+
+      cy.undoRedo().do("changeStateOrInfoBox", param);
+    }
+    else {
+      return elementUtilities.changeStateOrInfoBox(nodes, index, value, type);
+    }
+
+    cy.style().update();
+  };
+
+  // Add a new state or info box to given nodes.
+  // The box is represented by the parameter obj.
+  // Considers undoable option.
+  mainUtilities.addStateOrInfoBox = function(nodes, obj) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (!options.undoable) {
+      elementUtilities.addStateOrInfoBox(nodes, obj);
+    }
+    else {
+      var param = {
+        obj: obj,
+        nodes: nodes
+      };
+
+      cy.undoRedo().do("addStateOrInfoBox", param);
+    }
+
+    cy.style().update();
+  };
+
+  // Remove the state or info boxes of the given nodes at given index.
+  // Considers undoable option.
+  mainUtilities.removeStateOrInfoBox = function(nodes, index) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (!options.undoable) {
+      elementUtilities.removeStateOrInfoBox(nodes, {index: index});
+    }
+    else {
+      var param = {
+        locationObj: {index: index},
+        nodes: nodes
+      };
+
+      cy.undoRedo().do("removeStateOrInfoBox", param);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Set multimer status of the given nodes to the given status.
+   * Considers undoable option.
+   */
+  mainUtilities.setMultimerStatus = function(nodes, status) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (options.undoable) {
+      var param = {
+        status: status,
+        nodes: nodes,
+        firstTime: true
+      };
+
+      cy.undoRedo().do("setMultimerStatus", param);
+    }
+    else {
+      elementUtilities.setMultimerStatus(nodes, status);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Set clone marker status of given nodes to the given status.
+   * Considers undoable option.
+   */
+  mainUtilities.setCloneMarkerStatus = function(nodes, status) {
+    if (nodes.length === 0) {
+      return;
+    }
+
+    if (options.undoable) {
+      var param = {
+        status: status,
+        nodes: nodes,
+        firstTime: true
+      };
+
+      cy.undoRedo().do("setCloneMarkerStatus", param);
+    }
+    else {
+      elementUtilities.setCloneMarkerStatus(nodes, status);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Change style/css of given eles by setting getting property name to the given given value/values (Note that valueMap parameter may be
+   * a single string or an id to value map). Considers undoable option.
+   */
+  mainUtilities.changeCss = function(eles, name, valueMap) {
+    if (eles.length === 0) {
+      return;
+    }
+
+    if (!options.undoable) {
+      elementUtilities.changeCss(eles, name, valueMap);
+    }
+    else {
+      var param = {
+        eles: eles,
+        valueMap: valueMap,
+        name: name
+      };
+
+      cy.undoRedo().do("changeCss", param);
+    }
+
+    cy.style().update();
+  };
+
+  /*
+   * Change data of given eles by setting getting property name to the given given value/values (Note that valueMap parameter may be
+   * a single string or an id to value map). Considers undoable option.
+   */
+  mainUtilities.changeData = function(eles, name, valueMap) {
+    if (eles.length === 0) {
+      return;
+    }
+
+    if (!options.undoable) {
+      elementUtilities.changeData(eles, name, valueMap);
+    }
+    else {
+      var param = {
+        eles: eles,
+        valueMap: valueMap,
+        name: name
+      };
+
+      cy.undoRedo().do("changeData", param);
+    }
+
+    cy.style().update();
+  };
+
+
+  /*
+   * Hides given eles (the ones which are selected) and perform given layout afterward. Layout parameter may be layout options
+   * or a function to call. Requires viewUtilities extension and considers undoable option.
+   */
+  mainUtilities.hideAndPerformLayout = function(eles, layoutparam) {
+      var nodes = eles.nodes(); // Ensure that nodes list just include nodes
+
+      var allNodes = cy.nodes(":visible");
+      var nodesToShow = chise.elementUtilities.extendRemainingNodes(nodes, allNodes);
+      var nodesToHide = allNodes.not(nodesToShow);
+
+      if (nodesToHide.length === 0) {
+          return;
+      }
+
+      if (!options.undoable) {
+
+          var nodesWithHiddenNeighbor = cy.edges(":hidden").connectedNodes(':visible');
+          chise.thinBorder(nodesWithHiddenNeighbor);
+          elementUtilities.hideAndPerformLayout(nodesToHide, layoutparam);
+          var nodesWithHiddenNeighbor = cy.edges(":hidden").connectedNodes(':visible');
+          chise.thickenBorder(nodesWithHiddenNeighbor);
+      }
+      else {
+          var param = {
+              eles: nodesToHide,
+              layoutparam: layoutparam,
+              firstTime: true
+          };
+
+          var ur = cy.undoRedo();
+          ur.action("thickenBorder", chise.thickenBorder, chise.thinBorder);
+          ur.action("thinBorder", chise.thinBorder, chise.thickenBorder);
+
+          var actions = [];
+          var nodesWithHiddenNeighbor = cy.edges(":hidden").connectedNodes().intersection(nodesToHide);
+          actions.push({name: "thinBorder", param: nodesWithHiddenNeighbor});
+          actions.push({name: "hideAndPerformLayout", param: param});
+          nodesWithHiddenNeighbor = nodesToHide.neighborhood(":visible").nodes().difference(nodesToHide).difference(cy.nodes("[thickBorder]"));
+          actions.push({name: "thickenBorder", param: nodesWithHiddenNeighbor});
+          cy.undoRedo().do("batch", actions);
+      }
+  };
+
+  /*
+   * Shows all elements (the ones which are hidden if any) and perform given layout afterward. Layout parameter may be layout options
+   * or a function to call. Requires viewUtilities extension and considers undoable option.
+   */
+  mainUtilities.showAllAndPerformLayout = function(layoutparam) {
+    var hiddenEles = cy.elements(':hidden');
+    if (hiddenEles.length === 0) {
+      return;
+    }
+    if (!options.undoable) {
+      var nodesWithHiddenNeighbor = cy.edges(":hidden").connectedNodes(':visible');
+      chise.thinBorder(nodesWithHiddenNeighbor);
+      elementUtilities.showAndPerformLayout(hiddenEles, layoutparam);
+    }
+    else {
+      var param = {
+        eles: hiddenEles,
+        layoutparam: layoutparam,
+        firstTime: true
+      };
+
+      var ur = cy.undoRedo();
+      ur.action("thickenBorder", chise.thickenBorder, chise.thinBorder);
+      ur.action("thinBorder", chise.thinBorder, chise.thickenBorder);
+
+      var actions = [];
+      var nodesWithHiddenNeighbor = cy.nodes("[thickBorder]");
+      actions.push({name: "thinBorder", param: nodesWithHiddenNeighbor});
+      actions.push({name: "showAndPerformLayout", param: param});
+      cy.undoRedo().do("batch", actions);
+    }
+  };
+
+  /*
+   * Unhide given eles (the ones which are hidden if any) and perform given layout afterward. Layout parameter may be layout options
+   * or a function to call. Requires viewUtilities extension and considers undoable option.
+   */
+  mainUtilities.showAndPerformLayout = function(mainEle, eles, layoutparam) {
+      var hiddenEles = eles.filter(':hidden');
+      if (hiddenEles.length === 0) {
+          return;
+      }
+      mainUtilities.closeUpElements(mainEle, hiddenEles.nodes());
+      if (!options.undoable) {
+          var nodesWithHiddenNeighbor = cy.edges(":hidden").connectedNodes(':visible');
+          chise.thinBorder(nodesWithHiddenNeighbor);
+          elementUtilities.showAndPerformLayout(hiddenEles, layoutparam);
+          var nodesWithHiddenNeighbor = cy.edges(":hidden").connectedNodes(':visible');
+          chise.thickenBorder(nodesWithHiddenNeighbor);
+      }
+      else {
+          var param = {
+              eles: hiddenEles,
+              layoutparam: layoutparam,
+              firstTime: true
+          };
+
+          var ur = cy.undoRedo();
+          ur.action("thickenBorder", chise.thickenBorder, chise.thinBorder);
+          ur.action("thinBorder", chise.thinBorder, chise.thickenBorder);
+
+          var actions = [];
+          var nodesToThinBorder = (hiddenEles.neighborhood(":visible").nodes("[thickBorder]"))
+                  .difference(cy.edges(":hidden").difference(hiddenEles.edges().union(hiddenEles.nodes().connectedEdges())).connectedNodes());
+          actions.push({name: "thinBorder", param: nodesToThinBorder});
+          actions.push({name: "showAndPerformLayout", param: param});
+          var nodesToThickenBorder = hiddenEles.nodes().edgesWith(cy.nodes(":hidden").difference(hiddenEles.nodes()))
+  	            .connectedNodes().intersection(hiddenEles.nodes());
+          actions.push({name: "thickenBorder", param: nodesToThickenBorder});
+          cy.undoRedo().do("batch", actions);
+      }
+  };
+
+  /*
+  * Takes the hidden elements close to the nodes whose neighbors will be shown
+  * */
+  mainUtilities.closeUpElements = function(mainEle, hiddenEles) {
+      var leftX = Number.MAX_VALUE;
+      var rightX = Number.MIN_VALUE;
+      var topY = Number.MAX_VALUE;
+      var bottomY = Number.MIN_VALUE;
+      // Check the x and y limits of all hidden elements and store them in the variables above
+      hiddenEles.forEach(function( ele ){
+          if (ele.data('class') != 'compartment' &&  ele.data('class') != 'complex')
+          {
+              var halfWidth = ele.outerWidth()/2;
+              var halfHeight = ele.outerHeight()/2;
+              if (ele.position("x") - halfWidth < leftX)
+                  leftX = ele.position("x") - halfWidth;
+              if (ele.position("x") + halfWidth > rightX)
+                  rightX = ele.position("x") + halfWidth;
+              if (ele.position("y") - halfHeight < topY)
+                  topY = ele.position("y") - halfHeight;
+              if (ele.position("y") + halfHeight > topY)
+                  bottomY = ele.position("y") + halfHeight;
+          }
+      });
+
+      //The coordinates of the old center containing the hidden nodes
+      var oldCenterX = (leftX + rightX)/2;
+      var oldCenterY = (topY + bottomY)/2;
+
+      //Here we calculate two parameters which define the area in which the hidden elements are placed initially
+      var minHorizontalParam = mainEle.outerWidth()/2 + (rightX - leftX)/2;
+      var maxHorizontalParam = mainEle.outerWidth() + (rightX - leftX)/2;
+      var minVerticalParam = mainEle.outerHeight()/2 + (bottomY - topY)/2;
+      var maxVerticalParam = mainEle.outerHeight() + (bottomY - topY)/2;
+
+      //Quadrants is an object of the form {first:"obtained", second:"free", third:"free", fourth:"obtained"}
+      // which holds which quadrant are free (that's where hidden nodes will be brought)
+      var quadrants = mainUtilities.checkOccupiedQuadrants(mainEle, hiddenEles);
+      var freeQuadrants = [];
+      for (var property in quadrants) {
+          if (quadrants[property] === "free")
+              freeQuadrants.push(property);
+      }
+
+      //Can take values 1 and -1 and are used to place the hidden nodes in the random quadrant
+      var horizontalMult;
+      var verticalMult;
+      if (freeQuadrants.length > 0)
+      {
+        if (freeQuadrants.length === 3)
+        {
+          if (freeQuadrants.includes('first') && freeQuadrants.includes('second') && freeQuadrants.includes('third'))
+          {
+            horizontalMult = -1;
+            verticalMult = -1;
+          }
+          else if (freeQuadrants.includes('first') && freeQuadrants.includes('second') && freeQuadrants.includes('fourth'))
+          {
+            horizontalMult = 1;
+            verticalMult = -1;
+          }
+          else if (freeQuadrants.includes('first') && freeQuadrants.includes('third') && freeQuadrants.includes('fourth'))
+          {
+            horizontalMult = 1;
+            verticalMult = 1;
+          }
+          else if (freeQuadrants.includes('second') && freeQuadrants.includes('third') && freeQuadrants.includes('fourth'))
+          {
+            horizontalMult = -1;
+            verticalMult = 1;
+          }
+        }
+        else
+        {
+          //Randomly picks one quadrant from the free quadrants
+          var randomQuadrant = freeQuadrants[Math.floor(Math.random()*freeQuadrants.length)];
+
+          if (randomQuadrant === "first") {
+              horizontalMult = 1;
+              verticalMult = -1;
+          }
+          else if (randomQuadrant === "second") {
+              horizontalMult = -1;
+              verticalMult = -1;
+          }
+          else if (randomQuadrant === "third") {
+              horizontalMult = -1;
+              verticalMult = 1;
+          }
+          else if (randomQuadrant === "fourth") {
+              horizontalMult = 1;
+              verticalMult = 1;
+          }
+        }
+      }
+      else
+      {
+          horizontalMult = 0;
+          verticalMult = 0;
+      }
+      // If the horizontalMult is 0 it means that no quadrant is free, so we randomly choose a quadrant
+      var horizontalParam = mainUtilities.generateRandom(minHorizontalParam,maxHorizontalParam,horizontalMult);
+      var verticalParam = mainUtilities.generateRandom(minVerticalParam,maxVerticalParam,verticalMult);
+
+      //The coordinates of the center where the hidden nodes will be transfered
+      var newCenterX = mainEle.position("x") + horizontalParam;
+      var newCenterY = mainEle.position("y") + verticalParam;
+
+      var xdiff = newCenterX - oldCenterX;
+      var ydiff = newCenterY - oldCenterY;
+
+      //Change the position of hidden elements
+      hiddenEles.forEach(function( ele ){
+          var newx = ele.position("x") + xdiff;
+          var newy = ele.position("y") + ydiff;
+          ele.position("x", newx);
+          ele.position("y",newy);
+      });
+  };
+
+  /*
+   * Generates a number between 2 nr and multimplies it with 1 or -1
+   * */
+  mainUtilities.generateRandom = function(min, max, mult) {
+      var val = [-1,1];
+      if (mult === 0)
+          mult = val[Math.floor(Math.random()*val.length)];
+      return (Math.floor(Math.random() * (max - min + 1)) + min) * mult;
+  };
+
+  /*
+   * This function makes sure that the random number lies in free quadrant
+   * */
+  mainUtilities.checkOccupiedQuadrants = function(mainEle, hiddenEles) {
+      if (chise.getMapType() == 'PD')
+      {
+        var visibleNeighborEles = mainEle.neighborhood().difference(hiddenEles).nodes();
+        var visibleNeighborsOfNeighbors = visibleNeighborEles.neighborhood().difference(hiddenEles).difference(mainEle).nodes();
+        var visibleEles = visibleNeighborEles.union(visibleNeighborsOfNeighbors);
+      }
+      else
+        var visibleEles = mainEle.neighborhood().difference(hiddenEles).nodes();
+      var occupiedQuadrants = {first:"free", second:"free", third:"free", fourth:"free"};
+
+      visibleEles.forEach(function( ele ){
+          if (ele.data('class') != 'compartment' &&  ele.data('class') != 'complex')
+          {
+              if (ele.position("x") < mainEle.position("x") && ele.position("y") < mainEle.position("y"))
+                  occupiedQuadrants.second = "occupied";
+              else if (ele.position("x") > mainEle.position("x") && ele.position("y") < mainEle.position("y"))
+                  occupiedQuadrants.first = "occupied";
+              else if (ele.position("x") < mainEle.position("x") && ele.position("y") > mainEle.position("y"))
+                  occupiedQuadrants.third = "occupied";
+              else if (ele.position("x") > mainEle.position("x") && ele.position("y") > mainEle.position("y"))
+                  occupiedQuadrants.fourth = "occupied";
+          }
+      });
+      return occupiedQuadrants;
+  };
+
+  // Overrides highlightProcesses from SBGNVIZ - do not highlight any nodes when the map type is AF
+  mainUtilities.highlightProcesses = function(_nodes) {
+    if (elementUtilities.getMapType() == "AF")
+      return;
+    libs.sbgnviz.highlightProcesses(_nodes);
+  };
+
+  /**
+   * Resets map type to undefined
+   */
+  mainUtilities.resetMapType = function(){
+    elementUtilities.resetMapType();
+  };
+
+  /**
+   * return : map type
+   */
+  mainUtilities.getMapType = function(){
+    return elementUtilities.getMapType();
+  };
+
+  return mainUtilities;
+};
