--- src/utilities/option-utilities.js
@@ -1,63 +1,0 @@
-/*
- *  Extend default options and get current options by using this file 
- */
-
-// default options
-var defaults = {
-  // The path of core library images when sbgnviz is required from npm and the index html 
-  // file and node_modules are under the same folder then using the default value is fine
-  imgPath: 'node_modules/sbgnviz/src/img',
-  // Whether to fit labels to nodes
-  fitLabelsToNodes: function () {
-    return false;
-  },
-  fitLabelsToInfoboxes: function () {
-    return false;
-  },
-  // dynamic label size it may be 'small', 'regular', 'large'
-  dynamicLabelSize: function () {
-    return 'regular';
-  },
-  // percentage used to calculate compound paddings
-  compoundPadding: function () {
-    return 10;
-  },
-  // Whether to adjust node label font size automatically.
-  // If this option return false do not adjust label sizes according to node height uses node.data('font-size')
-  // instead of doing it.
-  adjustNodeLabelFontSizeAutomatically: function() {
-    return true;
-  },
-  // The selector of the component containing the sbgn network
-  networkContainerSelector: '#sbgn-network-container',
-  // Whether the actions are undoable, requires cytoscape-undo-redo extension
-  undoable: true,
-  // Whether to have undoable drag feature in undo/redo extension. This options will be passed to undo/redo extension
-  undoableDrag: true
-};
-
-var optionUtilities = function () {
-};
-
-// Extend the defaults options with the user options
-optionUtilities.extendOptions = function (options) {
-  var result = {};
-
-  for (var prop in defaults) {
-    result[prop] = defaults[prop];
-  }
-  
-  for (var prop in options) {
-    result[prop] = options[prop];
-  }
-
-  optionUtilities.options = result;
-
-  return options;
-};
-
-optionUtilities.getOptions = function () {
-  return optionUtilities.options;
-};
-
-module.exports = optionUtilities;
