+++ src/utilities/sbgnviz-instance-utilities-factory.js
@@ -1,0 +1,23 @@
+var libs = require('./lib-utilities').getLibs();
+
+module.exports = function () {
+
+  var instance;
+
+  function sbgnvizInstanceUtilities (options) {
+
+    instance = libs.sbgnviz(options);
+
+    return instance;
+  }
+
+  sbgnvizInstanceUtilities.getInstance = function () {
+    return instance;
+  }
+
+  sbgnvizInstanceUtilities.getCy = function () {
+    return this.getInstance().getCy();
+  }
+
+  return sbgnvizInstanceUtilities;
+};
