--- src/utilities/undo-redo-action-functions.js
@@ -1,480 +1,0 @@
-// Extends sbgnviz.undoRedoActionFunctions
-var libs = require('./lib-utilities').getLibs();
-var sbgnviz = libs.sbgnviz;
-var undoRedoActionFunctions = sbgnviz.undoRedoActionFunctions;
-var elementUtilities = require('./element-utilities');
-
-// Section Start
-// add/remove action functions
-
-undoRedoActionFunctions.addNode = function (param) {
-  var result;
-  if (param.firstTime) {
-    var newNode = param.newNode;
-    result = elementUtilities.addNode(newNode.x, newNode.y, newNode.class, newNode.id, newNode.parent, newNode.visibility);
-  }
-  else {
-    result = elementUtilities.restoreEles(param);
-  }
-
-  return {
-    eles: result
-  };
-};
-
-undoRedoActionFunctions.addEdge = function (param) {
-  var result;
-  if (param.firstTime) {
-    var newEdge = param.newEdge;
-    result = elementUtilities.addEdge(newEdge.source, newEdge.target, newEdge.class, newEdge.id, newEdge.visibility);
-  }
-  else {
-    result = elementUtilities.restoreEles(param);
-  }
-
-  return {
-    eles: result
-  };
-};
-
-undoRedoActionFunctions.addProcessWithConvenientEdges = function(param) {
-  var result;
-  if (param.firstTime) {
-    result = elementUtilities.addProcessWithConvenientEdges(param.source, param.target, param.processType);
-  }
-  else {
-    result = elementUtilities.restoreEles(param);
-  }
-
-  return {
-    eles: result
-  };
-};
-
-undoRedoActionFunctions.createCompoundForGivenNodes = function (param) {
-  var result = {};
-
-  if (param.firstTime) {
-    // Nodes to make compound, their descendants and edges connected to them will be removed during createCompoundForGivenNodes operation
-    // (internally by eles.move() operation), so mark them as removed eles for undo operation.
-    var nodesToMakeCompound = param.nodesToMakeCompound;
-    var removedEles = nodesToMakeCompound.union(nodesToMakeCompound.descendants());
-    removedEles = removedEles.union(removedEles.connectedEdges());
-    result.removedEles = removedEles;
-    // Assume that all nodes to make compound have the same parent
-    var oldParentId = nodesToMakeCompound[0].data("parent");
-    // The parent of new compound will be the old parent of the nodes to make compound
-    // New eles includes new compound and the moved eles and will be used in undo operation.
-    result.newEles = elementUtilities.createCompoundForGivenNodes(nodesToMakeCompound, param.compoundType);
-  }
-  else {
-    result.removedEles = param.newEles.remove();
-    result.newEles = param.removedEles.restore();
-    elementUtilities.maintainPointer(result.newEles);
-  }
-
-  return result;
-};
-
-// Section End
-// add/remove action functions
-
-// Section Start
-// easy creation action functions
-
-undoRedoActionFunctions.createTemplateReaction = function (param) {
-  var firstTime = param.firstTime;
-  var eles;
-
-  if (firstTime) {
-    eles = elementUtilities.createTemplateReaction(param.templateType, param.macromoleculeList, param.complexName, param.processPosition, param.tilingPaddingVertical, param.tilingPaddingHorizontal, param.edgeLength)
-  }
-  else {
-    eles = param;
-    cy.add(eles);
-    
-    cy.elements().unselect();
-    eles.select();
-  }
-
-  return {
-    eles: eles
-  };
-};
-
-// Section End
-// easy creation action functions
-
-// Section Start
-// general action functions
-
-undoRedoActionFunctions.getNodePositions = function () {
-  var positions = {};
-  var nodes = cy.nodes();
-  
-  nodes.each(function(ele, i) {
-    if(typeof ele === "number") {
-      ele = i;
-    }
-    
-    positions[ele.id()] = {
-      x: ele.position("x"),
-      y: ele.position("y")
-    };
-  });
-
-  return positions;
-};
-
-undoRedoActionFunctions.returnToPositions = function (positions) {
-  var currentPositions = {};
-  cy.nodes().positions(function (ele, i) {
-    if(typeof ele === "number") {
-      ele = i;
-    }
-    
-    currentPositions[ele.id()] = {
-      x: ele.position("x"),
-      y: ele.position("y")
-    };
-    
-    var pos = positions[ele.id()];
-    return {
-      x: pos.x,
-      y: pos.y
-    };
-  });
-
-  return currentPositions;
-};
-
-undoRedoActionFunctions.resizeNodes = function (param) {
-  var result = {
-    performOperation: true
-  };
-
-  var nodes = param.nodes;
-
-  result.sizeMap = {};
-  result.useAspectRatio = false;
-
-  for (var i = 0; i < nodes.length; i++) {
-    var node = nodes[i];
-    result.sizeMap[node.id()] = {
-      w: node.width(),
-      h: node.height()
-    };
-  }
-
-  result.nodes = nodes;
-
-  for (var i = 0; i < nodes.length; i++) {
-    var node = nodes[i];
-
-    if (param.performOperation) {
-      if (param.sizeMap) {
-        node.data("bbox").w = param.sizeMap[node.id()].w;
-        node.data("bbox").h = param.sizeMap[node.id()].h;
-      }
-      else {
-        elementUtilities.resizeNodes(param.nodes, param.width, param.height, param.useAspectRatio);
-      }
-    }
-  }
-
-  return result;
-};
-
-undoRedoActionFunctions.changeNodeLabel = function (param) {
-  var result = {
-  };
-  var nodes = param.nodes;
-  result.nodes = nodes;
-  result.label = {};
-
-  for (var i = 0; i < nodes.length; i++) {
-    var node = nodes[i];
-    result.label[node.id()] = node._private.data.label;
-  }
-
-  if (param.firstTime) {
-    nodes.data('label', param.label);
-  }
-  else {
-    for (var i = 0; i < nodes.length; i++) {
-      var node = nodes[i];
-      node._private.data.label = param.label[node.id()];
-    }
-  }
-
-  return result;
-};
-
-undoRedoActionFunctions.changeData = function (param) {
-  var result = {
-  };
-  var eles = param.eles;
-  result.name = param.name;
-  result.valueMap = {};
-  result.eles = eles;
-
-  for (var i = 0; i < eles.length; i++) {
-    var ele = eles[i];
-    result.valueMap[ele.id()] = ele.data(param.name);
-  }
-
-  elementUtilities.changeData(param.eles, param.name, param.valueMap);
-
-  return result;
-};
-
-undoRedoActionFunctions.changeCss = function (param) {
-  var result = {
-  };
-  var eles = param.eles;
-  result.name = param.name;
-  result.valueMap = {};
-  result.eles = eles;
-
-  for (var i = 0; i < eles.length; i++) {
-    var ele = eles[i];
-    result.valueMap[ele.id()] = ele.css(param.name);
-  }
-
-  elementUtilities.changeCss(param.eles, param.name, param.valueMap);
-
-  return result;
-};
-
-undoRedoActionFunctions.changeFontProperties = function (param) {
-  var result = {
-  };
-
-  var eles = param.eles;
-  result.data = {};
-  result.eles = eles;
-
-  for (var i = 0; i < eles.length; i++) {
-    var ele = eles[i];
-
-    result.data[ele.id()] = {};
-
-    var data = param.firstTime ? param.data : param.data[ele.id()];
-
-    for (var prop in data) {
-      result.data[ele.id()][prop] = ele.data(prop);
-    }
-  }
-
-  if (param.firstTime) {
-    elementUtilities.changeFontProperties(eles, data);
-  }
-  else {
-    for (var i = 0; i < eles.length; i++) {
-      var ele = eles[i];
-      
-      elementUtilities.changeFontProperties(ele, data);
-    }
-  }
-
-  return result;
-};
-
-/*
- * Show eles and perform layout.
- */
-undoRedoActionFunctions.showAndPerformLayout = function (param) {
-  var eles = param.eles;
-
-  var result = {};
-  result.positions = undoRedoActionFunctions.getNodePositions();
-  
-  if (param.firstTime) {
-    result.eles = elementUtilities.showAndPerformLayout(param.eles, param.layoutparam);
-  }
-  else {
-    result.eles = cy.viewUtilities().show(eles); // Show given eles
-    undoRedoActionFunctions.returnToPositions(param.positions);
-  }
-
-  return result;
-};
-
-undoRedoActionFunctions.undoShowAndPerformLayout = function (param) {
-  var eles = param.eles;
-
-  var result = {};
-  result.positions = undoRedoActionFunctions.getNodePositions();
-  result.eles = cy.viewUtilities().hide(eles); // Hide previously unhidden eles;
-
-  undoRedoActionFunctions.returnToPositions(param.positions);
-
-  return result;
-};
-
-/*
- * Hide eles and perform layout.
- */
-undoRedoActionFunctions.hideAndPerformLayout = function (param) {
-    var eles = param.eles;
-
-    var result = {};
-    result.positions = undoRedoActionFunctions.getNodePositions();
-
-    if (param.firstTime) {
-        result.eles = elementUtilities.hideAndPerformLayout(param.eles, param.layoutparam);
-    }
-    else {
-        result.eles = cy.viewUtilities().hide(eles); // Hide given eles
-        undoRedoActionFunctions.returnToPositions(param.positions);
-    }
-
-    return result;
-};
-
-undoRedoActionFunctions.undoHideAndPerformLayout = function (param) {
-    var eles = param.eles;
-
-    var result = {};
-    result.positions = undoRedoActionFunctions.getNodePositions();
-    result.eles = cy.viewUtilities().show(eles); // Show previously hidden eles
-
-    undoRedoActionFunctions.returnToPositions(param.positions);
-
-    return result;
-};
-
-// Section End
-// general action functions
-
-// Section Start
-// sbgn action functions
-
-undoRedoActionFunctions.changeStateOrInfoBox = function (param) {
-  var result = {
-  };
-  result.type = param.type;
-  result.nodes = param.nodes;
-  result.index = param.index;
-
-  result.value = elementUtilities.changeStateOrInfoBox(param.nodes, param.index, param.value, param.type);
-
-  cy.forceRender();
-
-  return result;
-};
-
-undoRedoActionFunctions.addStateOrInfoBox = function (param) {
-  var obj = param.obj;
-  var nodes = param.nodes;
-
-  var locationObj = elementUtilities.addStateOrInfoBox(nodes, obj);
-
-  cy.forceRender();
-
-  var result = {
-    nodes: nodes,
-    locationObj: locationObj,
-    obj: obj
-  };
-  return result;
-};
-
-undoRedoActionFunctions.removeStateOrInfoBox = function (param) {
-  var locationObj = param.locationObj;
-  var nodes = param.nodes;
-
-  var obj = elementUtilities.removeStateOrInfoBox(nodes, locationObj);
-
-  cy.forceRender();
-
-  var result = {
-    nodes: nodes,
-    obj: obj
-  };
-  return result;
-};
-
-undoRedoActionFunctions.setMultimerStatus = function (param) {
-  var firstTime = param.firstTime;
-  var nodes = param.nodes;
-  var status = param.status;
-  var resultStatus = {};
-
-  for (var i = 0; i < nodes.length; i++) {
-    var node = nodes[i];
-    var isMultimer = node.data('class').endsWith(' multimer');
-
-    resultStatus[node.id()] = isMultimer;
-  }
-
-  // If this is the first time change the status of all nodes at once.
-  // If not change status of each seperately to the values mapped to their id.
-  if (firstTime) {
-    elementUtilities.setMultimerStatus(nodes, status);
-  }
-  else {
-    for (var i = 0; i < nodes.length; i++) {
-      var node = nodes[i];
-      elementUtilities.setMultimerStatus(node, status[node.id()]);
-    }
-  }
-
-//  if (!firstTime && _.isEqual(nodes, cy.nodes(':selected'))) {
-//    $('#inspector-is-multimer').attr("checked", !$('#inspector-is-multimer').attr("checked"));
-//  }
-
-  var result = {
-    status: resultStatus,
-    nodes: nodes
-  };
-
-  return result;
-};
-
-undoRedoActionFunctions.setCloneMarkerStatus = function (param) {
-  var nodes = param.nodes;
-  var status = param.status;
-  var firstTime = param.firstTime;
-  var resultStatus = {};
-
-  for (var i = 0; i < nodes.length; i++) {
-    var node = nodes[i];
-    resultStatus[node.id()] = node.data('clonemarker');
-    var currentStatus = firstTime ? status : status[node.id()];
-    elementUtilities.setCloneMarkerStatus(node, currentStatus);
-  }
-
-//  if (!firstTime && _.isEqual(nodes, cy.nodes(':selected'))) {
-//    $('#inspector-is-clone-marker').attr("checked", !$('#inspector-is-clone-marker').attr("checked"));
-//  }
-
-  var result = {
-    status: resultStatus,
-    nodes: nodes
-  };
-
-  return result;
-};
-
-// param: {class: sbgnclass, name: propertyName, value: value}
-undoRedoActionFunctions.setDefaultProperty = function (param) {
-  var sbgnclass = param.class;
-  var name = param.name;
-  var value = param.value;
-  var classDefaults = elementUtilities.defaultProperties[sbgnclass];
-  var result = {
-    class: sbgnclass,
-    name: name,
-    value: classDefaults.hasOwnProperty(name) ? classDefaults[name] : undefined
-  };
-
-  classDefaults[name] = value;
-
-  return result;
-};
-
-// Section End
-// sbgn action functions
-
-module.exports = undoRedoActionFunctions;
